fix/cardsComponents
