+++ src/styles/components/tags/tags.sass
@@ -19,6 +19,39 @@
         line-height: 24px
 
 
+.tag-inative
+    background: variables.$color-grey-4
+    p
+        text-transform: capitalize
+        color: variables.$color-white-fixed
+        font-family: 'Inter'
+        font-style: normal
+        font-weight: 500
+        font-size: 14px
+        line-height: 24px
+        padding: 0px 8px
+
+
+.tag-ative
+    background: variables.$color-brand-1
+    p
+        text-transform: capitalize
+        color: variables.$color-white-fixed
+        font-family: 'Inter'
+        font-style: normal
+        font-weight: 500
+        font-size: 14px
+        line-height: 24px
+        padding: 0px 8px
+
+.tag-cash
+    background-color: variables.$color-random-7
+    right: 0px
+    svg
+        color: variables.$color-white-fixed
+        height: 27px
+        width: 15px
+
 .div-elipsis
     display: flex
     align-items: center
